# Put Laser abs(a) in separate Field