In this PR the absolute value of the laser envelope is stored in a separate field (inside the fields array). This has the benefit that when particles gather abs(a), they need to read only one value instead of real and imaginary part, resulting in a small performance improvement in plasma deposition and push if a laser is used. Additionally, this makes it easier to support multiple laser grids, independent laser resolution and laser with MR in the future.
A wrong indentation in the plasma pusher was fixed in this PR.


New
```
TinyProfiler total time across processes [min...avg...max]: 106.8 ... 112.2 ... 113.2

--------------------------------------------------------------------------------------------------
Name                                               NCalls  Excl. Min  Excl. Avg  Excl. Max   Max %
--------------------------------------------------------------------------------------------------
hpmg::MultiGrid::solve2()                            2000      21.14      22.04      23.42  20.68%
hpmg::MultiGrid::solve1()                            2000      20.73      21.57       22.4  19.79%
AnyDST::Execute()                                   12000      17.12      17.51      18.57  16.40%
MultiBuffer::get_data()                              2000   0.002343      6.927      16.05  14.18%
Fields::ShiftSlices()                                2000      1.125      4.214      10.36   9.15%
ExplicitDeposition()                                 2000      8.752      8.928      9.262   8.18%
MultiLaser::ShiftLaserSlices()                       2000      4.706      6.346      9.045   7.99%
AdvancePlasmaParticles()                             2000      6.328      6.571      7.702   6.80%
main()                                                  1   0.001286     0.7075      5.556   4.91%
DepositCurrent_PlasmaParticleContainer()             2001      4.289      4.411      4.676   4.13%
...
MultiLaser::GetLaserAabs()                           2000     0.3423     0.3644     0.4101   0.36%
...
--------------------------------------------------------------------------------------------------

Device Memory Usage:
-------------------------------------------------------------------------------------------------------------------------------------
Name                                            Nalloc  Nfree  AvgMem min  AvgMem avg  AvgMem max  MaxMem min  MaxMem avg  MaxMem max
-------------------------------------------------------------------------------------------------------------------------------------
The_Arena::Initialize()                              8      8     274 KiB     887 KiB    1049 KiB      47 GiB      47 GiB      47 GiB
MultiBuffer::put_data()                          20868  20868       0   B    3599 MiB      18 GiB       0   B    7288 MiB      25 GiB
MultiLaser::InitData()                              32     32    1366 MiB    1506 MiB    1530 MiB    1539 MiB    1539 MiB    1539 MiB
Fields::AllocData()                                  8      8    1392 MiB    1462 MiB    1476 MiB    1476 MiB    1476 MiB    1476 MiB
...
-------------------------------------------------------------------------------------------------------------------------------------
```
Old
```
TinyProfiler total time across processes [min...avg...max]: 109.8 ... 115.3 ... 116.4

--------------------------------------------------------------------------------------------------
Name                                               NCalls  Excl. Min  Excl. Avg  Excl. Max   Max %
--------------------------------------------------------------------------------------------------
hpmg::MultiGrid::solve2()                            2000      21.12      22.02      24.01  20.63%
hpmg::MultiGrid::solve1()                            2000      20.58       21.6      22.39  19.24%
AnyDST::Execute()                                   12000      17.31      17.68      18.71  16.08%
MultiBuffer::get_data()                              2000   0.002295      8.204      16.98  14.58%
ExplicitDeposition()                                 2000      10.09      10.21      10.36   8.90%
MultiLaser::ShiftLaserSlices()                       2000      4.616      6.338      9.857   8.47%
Fields::ShiftSlices()                                2000       1.13      3.727      9.151   7.86%
AdvancePlasmaParticles()                             2000      7.136       7.33       8.22   7.06%
main()                                                  1   0.001292       0.76      5.974   5.13%
DepositCurrent_PlasmaParticleContainer()             2001      4.681      4.816      5.241   4.50%
...
--------------------------------------------------------------------------------------------------

Device Memory Usage:
-------------------------------------------------------------------------------------------------------------------------------------
Name                                            Nalloc  Nfree  AvgMem min  AvgMem avg  AvgMem max  MaxMem min  MaxMem avg  MaxMem max
-------------------------------------------------------------------------------------------------------------------------------------
The_Arena::Initialize()                              8      8     347 KiB     860 KiB     960 KiB      47 GiB      47 GiB      47 GiB
MultiBuffer::put_data()                          19960  19960     308 KiB    3931 MiB      19 GiB     256 MiB    7416 MiB      25 GiB
MultiLaser::InitData()                              32     32    1364 MiB    1506 MiB    1530 MiB    1539 MiB    1539 MiB    1539 MiB
Fields::AllocData()                                  8      8    1332 MiB    1399 MiB    1412 MiB    1412 MiB    1412 MiB    1412 MiB
...
-------------------------------------------------------------------------------------------------------------------------------------
```


- [ ] **Small enough** (< few 100s of lines), otherwise it should probably be split into smaller PRs
- [ ] **Tested** (describe the tests in the PR description)
- [ ] **Runs on GPU** (basic: the code compiles and run well with the new module)
- [ ] **Contains an automated test** (checksum and/or comparison with theory)
- [ ] **Documented**: all elements (classes and their members, functions, namespaces, etc.) are documented
- [ ] **Constified** (All that can be `const` is `const`)
- [ ] **Code is clean** (no unwanted comments, )
- [ ] **Style and code conventions** are respected at the bottom of https://github.com/Hi-PACE/hipace
- [ ] **Proper label and GitHub project**, if applicable